requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
decorator==4.4.2
defusedxml==0.7.1
Django==3.1.7
django-crispy-forms==1.11.2
django-etc==1.3.0
django-hitcount==1.3.3
django-js-asset==1.2.2
django-taggit==1.3.0
django-widget-tweaks==1.4.8
Faker==6.6.2
idna==2.10
intervaltree==3.1.0
Markdown==3.3.4
networkx==2.5
oauthlib==3.1.0
Pillow==8.1.2
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sortedcontainers==2.3.0
sqlparse==0.4.1
tag==0.5
text-unidecode==1.3
urllib3==1.26.4
django-debug-toolbar~=3.2
